+++ src/libs/animation/variant/Fade/index.ts
@@ -1,0 +1,3 @@
+export type { FadeProps } from './type'
+
+export { genAnimation as genFadeAnimation } from './gen'