+++ src/libs/animation/variant/type.ts
@@ -1,5 +1,7 @@
 import type { CSSProperties } from 'react'
 
+import type { AnimationProps } from './AnimationFactory'
+
 export type AnimationBaseProps = {
   delay?: CSSProperties['animationDelay']
   duration?: CSSProperties['animationDuration']
@@ -10,4 +12,8 @@ export type AnimationBaseProps = {
 export type GeneratedAnimationValue = {
   className: string
   style: object
+}
+
+export type AnimationArgs = {
+  animationProps?: AnimationProps
 }