+++ types/src/components/Typography/index.stories.d.ts
@@ -1,0 +1,19 @@
+import { Typography } from '.';
+import type { Meta, StoryObj } from '@storybook/react';
+declare const meta: Meta<typeof Typography>;
+export default meta;
+type Story = StoryObj<typeof Typography>;
+export declare const H1: Story;
+export declare const H2: Story;
+export declare const H3: Story;
+export declare const H4: Story;
+export declare const H5: Story;
+export declare const H6: Story;
+export declare const Paragraph: Story;
+export declare const BodyLarge: Story;
+export declare const BodyMedium: Story;
+export declare const BodySmall: Story;
+export declare const Backquote: Story;
+export declare const Strong: Story;
+export declare const ParagraphWithColor: Story;
+//# sourceMappingURL=index.stories.d.ts.map